Forked from Community Installer
Standing on the shoulders of giants!

Added initial prompt for Dragonchain node name and use this parameter throughout to create directories and build the node. This allows for multiple runs of the script to create unique nodes.

Added CPU limits to helm chart setup (for Raspberry Pi primarily, but also works natively OK)

Upgrade script not amended.